L3 Interface Config (#19521)
* L3 Interface Config

* Cleanup

* Fixes for review comments and some cleanup

* Fixing some line diffs

* Fix for uuid parsing

+++ nic/agent/dscagent/types/types.go
@@ -17,8 +17,8 @@ type DistributedServiceCardStatus struct {
 
 //DSCInterfaceIP captures the IP interface info of DSC interfaces
 type DSCInterfaceIP struct {
-	IfID          int8
-	DestPrefixLen int8
+	IfID          uint32
+	DestPrefixLen uint32
 	IPAddress     string
 	GatewayIP     string
 }
